refactor: move enforce_distribution and enforce_sorting into ensure_requirements submodules
After the EnforceDistribution/EnforceSorting rule structs were removed,
the helpers in enforce_distribution.rs and enforce_sorting/ have a single
caller — ensure_requirements/mod.rs. Move both into ensure_requirements/
to make that internal-implementation-detail relationship explicit, per
alamb's review comments on #21976.

The crate root still re-exports both modules so external test files in
datafusion/core/tests/physical_optimizer/ (enforce_distribution.rs,
enforce_sorting.rs, etc.) keep their existing public paths.
